backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
pydantic>=2.0.0
readability-lxml>=0.8.0
